Refactor timereport
Extract functions to a new module.
This makes it easier to write unit tests.
